+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 1.5.29
 - Update Forge to 2577 (minimum Forge version required is now 2555 for MC 1.12.2)
+- Fixed bug where MCMP multiparts were blocking RS network connections (raoulvdberge)
 
 ### 1.5.28
 - Fixed Writers not pushing energy (raoulvdberge)